Refactored CreateWorkout into container and component files
