+++ client/src/components/dashPage/index.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import Navbar from './navbar.jsx';
 import {Tabs, Tab} from 'material-ui/Tabs';
 
-import CreateWorkout from '../CreateWorkout.jsx';
+import CreateWorkoutContainer from '../../containers/CreateWorkoutContainer.jsx';
 
 const tabStyle = {
   defaultTab: {
@@ -54,7 +54,7 @@ class DashPage extends Component{
             </div>
           </Tab>
         </Tabs>
-        <CreateWorkout user_id={this.props.userInfo.id} REST_SERVER_URL='http://localhost:8000/api' API_ENDPOINT='/workouts/addWorkout' />
+        <CreateWorkoutContainer user_id={this.props.userInfo.id} />
       </div>
     );
   }
